refactor: fix typo: manger -> manager
