+++ config/crd/kustomization.yaml
@@ -1,5 +1,5 @@
 # This kustomization.yaml is not intended to be run by itself,
-# since it depends on manger name and namespace that are out of this kustomize package.
+# since it depends on manager name and namespace that are out of this kustomize package.
 # It should be run by config/default
 resources:
 - bases/risingwave.singularity-data.com_risingwaves.yaml
