+++ config/crd/patches/webhook_in_risingwaves.yaml
@@ -10,7 +10,7 @@ spec:
       clientConfig:
         service:
           namespace: system
-          name: webhook-manger
+          name: webhook-manager
           path: /convert
       conversionReviewVersions:
       - v1
