+++ pkg/manager/meta.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package manger
+package manager
 
 import (
 	"context"
